requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
python-multipart==0.0.26
redis==7.4.0
requests==2.33.1
chardet==7.4.1
psycopg2-binary==2.9.11
streamlit==1.56.0
